Fix crash related to KnowItAll (#220)
**What kind of change does this PR introduce?** (Bug fix, feature, docs
update, ...)
Replaces the `KnowItAll` with a custom class which can be correctly
instantiated on the client.

**What is the current behavior?** (You can also link to an open issue
here)
Opening a workbench with an arcane recipe on a multiplayer server
crashes the game.
https://github.com/GTNewHorizons/GT-New-Horizons-Modpack/issues/19522

**What is the new behavior (if this is a feature change)?**
No longer crashes, and the `KnowItAll` is properly disposed of when the
world is closed.

**Does this PR introduce a breaking change?**
No, except if any PRs are relying on the old location of the
`KnowItAll`.