[Bugfix] Fix research error localization. (#258)
**What kind of change does this PR introduce?** (Bug fix, feature, docs
update, ...)
Bug fix-ish - fixes an issue with the research errors.

**What is the current behavior?** (You can also link to an open issue
here)
The research errors for the Crucible and the Runic Matrix were being
created on the server, which would cause the names of the categories and
the research to be translated on the server (to English, rather than
whatever the client's language is.)

**What is the new behavior (if this is a feature change)?**
Properly constructs the chat components to translate the substituted
parts on the client

**Does this PR introduce a breaking change?**
No

**Other information**:
This change doesn't do all that much good right now, considering that
the error message itself hasn't been translated to any other language
yet.