Allow "Single Wand Replacement" to work when the "missing research error in Arcane Workbench" feature is enabled. (#280)
**What kind of change does this PR introduce?** (Bug fix, feature, docs
update, ...)
Bug fix - the "missing research error" feature was preventing the
"single wand replacement" GUI changes from activating.

**What is the current behavior?** (You can also link to an open issue
here)
See above - when `notifyMissingResearchWorkbench` was set to true, the
Arcane Workbench GUI wouldn't show the vis cost & the research
requirements of the recipe unless there was a wand in the wand slot.

**What is the new behavior (if this is a feature change)?**
The mix-ins now target more logical injection points and use less
restrictive logic in order to work together.

**Does this PR introduce a breaking change?**
No

**Other information**:
Tested functional in dev & obfuscated environments.

New functionality:
- Trying to swap wand parts without the required research will show the
error text in the workbench when there's no wand in the wand slot.
- Trying to swap wand parts without a wand in the wand slot will show
the vis price of the swap. (even when `notifyMissingResearchWorkbench`
is turned on.)