Added extra security checks to Arcane Keys (#336)
**What kind of change does this PR introduce?** (Bug fix, feature, docs
update, ...)
Bug fix - keys no longer open doors that they aren't supposed to.

**What is the current behavior?** (You can also link to an open issue
here)
Keys would only check the coordinates of the block being unlocked.
This means one could make a key from a door in another dimension or a
door that was later broken and replaced, and use that key to open the
door even though they don't have permissions for it.

**What is the new behavior (if this is a feature change)?**
In order to open a door:
- The dimension ID must match
- The player who made the key must still have the ability to grant keys
to the door.

This PR also causes the tooltip to display the dimension ID and the key
creator to be saved into the NBT data & displayed in the tooltip. Any
old keys without the extra tags will function insecurely, as they
previously would.

**Does this PR introduce a breaking change?**
No