NEI paste now updates the output slot of the Arcane Worktable. (#343)
**What kind of change does this PR introduce?** (Bug fix, feature, docs
update, ...)
Bug fix

**What is the current behavior?** (You can also link to an open issue
here)
Closes #337

**What is the new behavior (if this is a feature change)?**
Pasting in a recipe from NEI visually updates the output slot to show
the new recipe.

**Does this PR introduce a breaking change?**
No

**Other information**:
I have no idea how NEI manages to modify the inventory of the Arcane
Worktable without causing a single recipe update, and I really don't
know how to keep chasing the issue. As far as I can tell, `slotClick`
gets called on the client, but the slots it modifies never update the
inventory, even though `setInventorySlotContents` is the only way to
change the items in the slots, which should trigger the update function.