Arcane Worktable Recipe Caching & Forge Event Bridge (#352)
**What kind of change does this PR introduce?** (Bug fix, feature, docs
update, ...)
- Bug fix - closes #278
- Improvement - bulk crafting is now near-instant instead of taking half
a second.

**What is the current behavior?** (You can also link to an open issue
here)
See issue #278 
Shift-clicking a craft out of the arcane workbench window can take a
while to calculate.

**What is the new behavior (if this is a feature change)?**
- Arcane Workbenches now cache the last valid recipe (or if no valid
recipe was found), allowing the GUI to render much faster (not profiled,
but it's got to save a good 50% of the runtime.)
- Arcane Workbenches are now visibly faster (shift-clicking a stack of
arcane crafting recipes took a good half-second on my dev environment,
is now as fast as a Crafting Table.)
- Forge events now have a bit more compatibility with standard
crafting-table triggers.

**Does this PR introduce a breaking change?**
No