Fix NEI disconnecting Arcane Workbench GUIs, for real this time. (#353)
**What kind of change does this PR introduce?** (Bug fix, feature, docs
update, ...)
Bug fix

**What is the current behavior?** (You can also link to an open issue
here)
Closes #337

**What is the new behavior (if this is a feature change)?**
After a `Container` with multi-container enabled is opened, it will
check whether it is connected to the tile entity & reconnect if it
hasn't. (The Arcane Workbench will also make sure that other players'
changes to the workbench (made while the `Container` was disconnected)
are reflected by forcing an update of the output slot.)

**Does this PR introduce a breaking change?**
This PR reverts #343, since that PR didn't actually fix the underlying
issue.