Expand focus upgrade packet limit (#372)
**What kind of change does this PR introduce?** (Bug fix, feature, docs
update, ...)
Bug fix - closes #369 

**What is the current behavior?** (You can also link to an open issue
here)
`GuiFocalManipulator` uses a vanilla "enchantment button pressed" packet
to send the selected ID of the upgrade to the server. Unfortunately,
this packet was not designed to support huge amounts of buttons. While
`C11PacketEnchantItem` could store an integer, it would only write a
byte of that integer when serialized.

The result was that, when playing on multiplayer servers, focus upgrade
IDs greater than 127 would be received as negative numbers on the server
& would get ignored.

**What is the new behavior (if this is a feature change)?**
Instead of sending a vanilla `C11PacketEnchantItem`, we send a custom
`MessageExtendedEnchantItem` packet that can store and transmit an
entire integer. (Please do note that focus upgrades are limited to
shorts, and considering they're stored in an array, picking extra large
IDs causes extremely large arrays to get allocated.)

**Does this PR introduce a breaking change?**
No